Continuation of last commit.

+++ game/world/managers/objects/script/ScriptHandler.py
@@ -125,13 +125,12 @@ class ScriptHandler:
         if event.action3_script > 0:
             self.ooc_scripts.append(event.action3_script)
 
-        self.ooc_spawn_min_delay = event.event_param1 / 1000
-        self.ooc_spawn_max_delay = event.event_param2 / 1000
-        self.ooc_repeat_min_delay = event.event_param3 / 1000
-        self.ooc_repeat_max_delay = event.event_param4 / 1000
+        self.ooc_spawn_min_delay = int(event.event_param1 / 1000)
+        self.ooc_spawn_max_delay = int(event.event_param2 / 1000)
+        self.ooc_repeat_min_delay = int(event.event_param3 / 1000)
+        self.ooc_repeat_max_delay = int(event.event_param4 / 1000)
 
         script = WorldDatabaseManager.creature_ai_script_get_by_id(random.choice(self.ooc_scripts))
-
         if script:
             self.ooc_target = target
 
@@ -154,7 +153,6 @@ class ScriptHandler:
 
         if scripts:
             for script in scripts:
-
                 if script.condition_id > 0:
                     if not ConditionChecker.check_condition(script.condition_id, self.object, target):
                         continue
